Fixed cast of Modify size_difference to int64_t.

+++ src/maidsafe/private/chunk_actions/appendable_by_all_rules.cc
@@ -309,7 +309,8 @@ int ProcessModify<kAppendableByAll>(const std::string &name,
     }
   }
 
-  *size_difference = existing_content.size() - new_content->size();
+  *size_difference = static_cast<int64_t>(existing_content.size()) -
+                     new_content->size();
   return kSuccess;
 }
